+++ DataAccess/Repository/IRepository/IUserRepository.cs
@@ -10,7 +10,7 @@ namespace DataAccess.Repository.IRepository
     public interface IUserRepository
     {
         Task<List<User>> GetAll();
-        Task<User?> FindById(int id);
+        Task<User?> FindById(long id);
         Task<User?> FindByEmail(string email);
         Task Add(User user);
         Task Update(User user);
